monkey patches the upcoming meetup home section into a testimonials section
