+++ src/pages/home/index.md
@@ -12,7 +12,7 @@ links:
     linkURL: 'https://github.com/mexitalian'
   - linkText: Audiovisual
     linkURL: 'https://www.instagram.com/howtovj/'
-upcomingMeetupHeading: Upcoming Meetup
+upcomingMeetupHeading: Testimonials
 noUpcomingMeetupText: Details to be announced.
 mapsNote: Clicking the pin opens Google Maps in a new tab.
 callToActions:
@@ -22,9 +22,9 @@ callToActions:
     linkURL: /meetups
     subHeading: Look at what topics were presented at past meetups.
   secondCTA:
-    heading: Volunteer to Present
+    heading: Let's talk
     linkType: external
-    linkURL: 'mailto:contact@js-wakanda.org'
+    linkURL: 'mailto:hello@charlos.co.uk'
     subHeading: Want to present at an upcoming meetup? Contact us.
 seo:
   browserTitle: JS Wakanda
