+++ src/pages/index.js
@@ -3,129 +3,19 @@ import PropTypes from "prop-types";
 import { graphql } from "gatsby";
 import Helmet from "react-helmet";
 import isAfter from "date-fns/is_after";
-
 import Layout from "../components/Layout";
-import Map from "../components/Map";
-import HeadshotPlaceholder from "../img/headshot-placeholder.svg";
-import CustomLink from "../components/CustomLink";
+import HomePageTemplate from "../templates/home";
 import "../styles/home.scss";
 
 
-const Links = ({ links }) => (
-  <ul className="hero-links">
-    {links && links.map((link, index) => (
-        <li key={`hero-links-${index}`} className="hero-linkItem">
-          <a className="hero-link" href={link.linkURL}>
-            {link.linkText}
-          </a>
-        </li>
-      ))}
-  </ul>
-)
-
-const Header = ({ home }) => (
-  <section className="header">
-    <div className="header-container container">
-      {home.headerImage && <img className="header-image" src={home.headerImage.image} alt={home.headerImage.imageAlt} />}
-      <div className="header-tagline-container">
-        <h3 className="header-tagline">
-          <span className="header-taglinePart">{home.title}</span>
-        </h3>
-        <h4 className="header-subheading">{home.subtitle}</h4>
-      </div>
-      <Links links={home.links}/>
-    </div>
-  </section>
-);
-
-
-const Testimonials = ({ home, testimonals = null }) => {
-  const presenters = testimonals && testimonals.presenters;
-  const latitude = testimonals && parseFloat(testimonals.location.mapsLatitude);
-  const longitude = testimonals && parseFloat(testimonals.location.mapsLongitude);
-  return (
-    <section className="upcomingMeetup  section">
-      <div className="upcomingMeetup-container  container">
-        <h2 className="upcomingMeetup-title">{home.upcomingMeetupHeading}</h2>
-        {testimonals ? (
-          <>
-            <p className="upcomingMeetup-detail  upcomingMeetup-detail--location">
-              <span className="upcomingMeetup-detailLabel">Currently living in: </span>
-              {testimonals.location.name}
-            </p>
-            {presenters.length > 0 && (
-              <div className="upcomingMeetup-presenters">
-                {presenters.map(presenter => (
-                  <div className="upcomingMeetup-presenter" key={presenter.text}>
-                    <img
-                      className="upcomingMeetup-presenterImage"
-                      src={presenter.image ? presenter.image : HeadshotPlaceholder}
-                      alt={presenter.image ? presenter.name : "Default headshot placeholder"}
-                    />
-                    <span className="upcomingMeetup-presenterName">{presenter.name}</span>
-                    <span className="upcomingMeetup-presenterPresentationTitle">
-                      {presenter.presentationTitle}
-                    </span>
-                    <p className="upcomingMeetup-presenterDescription">{presenter.text}</p>
-                  </div>
-                ))}
-              </div>
-            )}
-            <p className="upcomingMeetup-mapNote">{home.mapsNote}</p>
-            <div className="upcomingMeetup-mapWrapper">
-              <Map
-                isMarkerShown
-                googleMapURL="https://maps.googleapis.com/maps/api/js?key=AIzaSyBTxauB_VWpo0_8hWELlE3pN59uuHzxD-8&v=3.exp&libraries=geometry,drawing,places"
-                loadingElement={<div style={{ height: `100%` }} />}
-                containerElement={<div style={{ height: `100%` }} />}
-                mapElement={<div style={{ height: `100%` }} />}
-                link={testimonals.location.mapsLink}
-                latitude={latitude}
-                longitude={longitude}
-              />
-            </div>
-          </>
-        ) : (
-          <p className="upcomingMeetup-detail">{home.noUpcomingMeetupText}</p>
-        )}
-      </div>
-    </section>
-  )
-};
-
-
-export const HomePageTemplate = ({ home, upcomingMeetup = null }) => {
-  return (
-    <>
-      <Header home={home} />
-      <Testimonials home={home} testimonals={upcomingMeetup} />
-      <section className="ctaBlock">
-        <CustomLink
-          linkType={home.callToActions.firstCTA.linkType}
-          linkURL={home.callToActions.firstCTA.linkURL}
-          className="ctaBlock-pattern  ctaBlock-pattern--first"
-        >
-          <div className="ctaBlock-cta">
-            <span className="ctaBlock-ctaHeading">{home.callToActions.firstCTA.heading}</span>
-            <p className="ctaBlock-ctaDescription">{home.callToActions.firstCTA.subHeading}</p>
-          </div>
-        </CustomLink>
-        <CustomLink
-          linkType={home.callToActions.secondCTA.linkType}
-          linkURL={home.callToActions.secondCTA.linkURL}
-          className="ctaBlock-pattern  ctaBlock-pattern--second"
-        >
-          <div className="ctaBlock-cta">
-            <span className="ctaBlock-ctaHeading">{home.callToActions.secondCTA.heading}</span>
-            <p className="ctaBlock-ctaDescription">{home.callToActions.secondCTA.subHeading}</p>
-          </div>
-        </CustomLink>
-      </section>
-    </>
-  );
-};
-
 class HomePage extends React.Component {
+  static propTypes = {
+    data: PropTypes.shape({
+      allMarkdownRemark: PropTypes.shape({
+        edges: PropTypes.array,
+      }),
+    }),
+  };
   render() {
     const { data } = this.props;
     const {
@@ -159,14 +49,6 @@ class HomePage extends React.Component {
   }
 }
 
-HomePage.propTypes = {
-  data: PropTypes.shape({
-    allMarkdownRemark: PropTypes.shape({
-      edges: PropTypes.array,
-    }),
-  }),
-};
-
 export default HomePage;
 
 export const pageQuery = graphql`
@@ -191,6 +73,10 @@ export const pageQuery = graphql`
               image
               text
               presentationTitle
+              links {
+                linkURL
+                linkText
+              }
             }
             location {
               mapsLatitude
